refactor(fixtures): the snapshot half moves out, under the 800-line ratchet
Measured against `origin/main` rather than against the previous commit,
which is the comparison the ratchet is really making and the one this
branch had not run: `kstrl/fixtures.py` is 792 lines on `1775301` and was
838 here, so `file_length_ratchet.py` reports CROSSED and exits 1. Per
commit it never fired, because each commit compares against the branch's
own HEAD and by then the file was already over.

This is the class CLAUDE.md records under "merging into a moving main":
a ratchet failure attached to no single hunk, present in neither the
branch's last commit nor `main`. It is also the standard this pull
request already set for itself twice, splitting `kstrl/dampener.py` at
927 lines and `tests/test_sense_dampener.py` at 855.

The seam is the one the ratchet's own text asks for. Running a fixture
and remembering what it produced last time are two jobs, and only the
second touches the snapshot directory: `save_snapshot` and
`check_snapshot_regression` move to `kstrl/fixtures_snapshot.py`, 752
lines and 111. No logic changes, and the two functions are byte-identical
to what they were.

`kstrl/fixtures.py` imports the new module to call both from
`check_fixtures`; the new module imports `Fixture` and `FixtureResult`
back under TYPE_CHECKING only, because a run-time import would be a
cycle and neither function constructs either type. Three test import
sites move with them, including the source string
`tests/test_atomicio.py` feeds to a subprocess.

`docs/atlas/atlas.json` still files both under `kstrl.fixtures`. It is a
generated artifact built at `ba6b182`, nothing in the repository
regenerates or validates it, and it is stale against `main` already, so
it is left rather than hand-edited.
